Update mpas global sum real subroutine
This commit updates the global sum real subroutine, cam_mpas_global_sum_real in
two ways. The first, is to be a function, rather than a subroutine with an out
argument and the second is that it removes the dminfo argument because there is
already a reference to dminfo within the cam_mpas_subdriver.

It also updates all of the calls to cam_mpas_global_sum_real.

In the furture, another commit could update cam_mpas_global_sum_real further by
adding a similar function to compute the global sum of a integer, and then add
a generic interface for the two.