Add dry air scaling nl option to cam_initfiles_nl
This commit removes the logical, mpas_scale_dry_air_mass namelist option and
replaces it a real-valued, scale_dry_air_mass nl option. Instead of having the
scale_dry_air_mass option in nhyd_model namelist group, it has been moved to
cam_initfiles_nl so that it can later be used by other dycores.

Currently, the namelist is only used by the MPAS dycore.

This commit also updates the calling logic in the mpas dycore so that
set_dry_mass is not called when running with simple physics and if the
scale_dry_air_mass value is 0.0.